web/uuidy/deploy/requirements.txt
aiogram==3.13.1
asyncpg==0.30.0
cachetools==5.5.0
PyJWT==2.9.0
PyJWT==2.9.0
quart==0.19.8
